+++ deepchem/models/tensorgraph/models/graph_conv.py
@@ -84,80 +84,3 @@ class GraphConvTensorGraph(TensorGraph):
       time2 = time.time()
       print("TIMING: model fitting took %0.3f s" % (time2 - time1))
       ############################################################## TIMING
-
-
-def graph_conv_model(batch_size, num_tasks):
-  model = GraphConvTensorGraph(batch_size=batch_size, use_queue=False)
-  atom_features = Input(shape=(None, 75))
-  model.add_feature(atom_features)
-
-  degree_slice = Input(shape=(None, 2), dtype=tf.int32)
-  model.add_feature(degree_slice)
-
-  membership = Input(shape=(None,), dtype=tf.int32)
-  model.add_feature(membership)
-
-  deg_adjs = []
-  for i in range(model.min_degree, model.max_degree + 1):
-    deg_adj = Input(shape=(None, i + 1), dtype=tf.int32)
-    model.add_feature(deg_adj)
-    deg_adjs.append(deg_adj)
-
-  gc1 = GraphConvLayer(64, activation_fn=tf.nn.relu)
-  model.add_layer(
-      gc1, parents=[atom_features, degree_slice, membership] + deg_adjs)
-
-  batch_norm1 = BatchNormLayer()
-  model.add_layer(batch_norm1, parents=[gc1])
-
-  gp1 = GraphPoolLayer()
-  model.add_layer(
-      gp1, parents=[batch_norm1, degree_slice, membership] + deg_adjs)
-
-  gc2 = GraphConvLayer(64, activation_fn=tf.nn.relu)
-  model.add_layer(gc2, parents=[gp1, degree_slice, membership] + deg_adjs)
-
-  batch_norm2 = BatchNormLayer()
-  model.add_layer(batch_norm2, parents=[gc2])
-
-  gp2 = GraphPoolLayer()
-  model.add_layer(
-      gp2, parents=[batch_norm2, degree_slice, membership] + deg_adjs)
-
-  dense = Dense(out_channels=128, activation_fn=None)
-  model.add_layer(dense, parents=[gp2])
-
-  batch_norm3 = BatchNormLayer()
-  model.add_layer(batch_norm3, parents=[dense])
-
-  gg1 = GraphGather(batch_size=batch_size, activation_fn=tf.nn.tanh)
-  model.add_layer(
-      gg1, parents=[batch_norm3, degree_slice, membership] + deg_adjs)
-
-  costs = []
-  for task in range(num_tasks):
-    classification = Dense(
-        out_channels=2, name="GUESS%s" % task, activation_fn=None)
-    model.add_layer(classification, parents=[gg1])
-
-    softmax = SoftMax(name="SOFTMAX%s" % task)
-    model.add_layer(softmax, parents=[classification])
-    model.add_output(softmax)
-
-    label = Input(shape=(None, 2), name="LABEL%s" % task)
-    model.add_label(label)
-
-    cost = SoftMaxCrossEntropy(name="COST%s" % task)
-    model.add_layer(cost, parents=[label, classification])
-    costs.append(cost)
-
-  entropy = Concat(name="ENT")
-  model.add_layer(entropy, parents=costs)
-
-  task_weights = Input(shape=(None, num_tasks), name="W")
-  model.add_task_weight(task_weights)
-
-  loss = WeightedError(name="ERROR")
-  model.add_layer(loss, parents=[entropy, task_weights])
-  model.set_loss(loss)
-  return model
